ASSERT_NOTHROW should rethrow
After an exception is caught, the result of the operation is null so the test can't really continue. So we throw an exception so be caught by the TEST macro. ASSERT_NOTHROW lets us determine which call threw an exception. TEST just tells us that some function threw an exception.