No more classes
Structs can have member functions so it's not that big of a deal. Maybe I'll make it possible for the user to implement classes if they really need virtual functions. But you need pointers to do that...

Pointers are strange. Many high level languages don't have pointers because "they are dangerous". Pointers are powerful. Inheritance and virtual functions are not actually that hard to implement in C. C has pointers.

I don't think it would be possible for the user of this language to implement this themselves if they need to. For that, I would need to expose pointers. Swift exposes pointers because pointers are needed by a systems language.

I don't want any undefined behaviour. Keeping pointers out is the best way to avoid this.