Use a few more enums in the AST
I can't decide if I should use ALL_CAPS for enums or not. I've been using ALL_CAPS but a lot of people say that ALL_CAPS is reserved for macros and I should just use lower_case instead. That's kind of annoying when creating a language because I need to put keywords in enums. public and private are keywords in C++ so I have to use pub and priv instead.