Goodbye Xcode!
CLion is a far, far, far superior C++ IDE. You only have to use it for 5 minutes to realise that.

Xcode project settings is a huge list of properties that I can set, whereas CMake is just the settings I want, in a file. Lovely!

The only reason I was using Xcode is for the most petty reason ever: syntax highlighting. When you've been looking at the same syntax highlighting theme for years, any other theme looks bad. I've chosen a nice theme that I hope I'll get used to.

CMake means that you don't need a Mac to build the project! Who's going to even consider using this project if they have to stuff around to get it to build on their machine?!