Dictionary literals start with [{
When we encounter a [ it could be an array or a dictionary. This is annoying to parse and might produce some bad error messages. When we encounter a [{ it must be a dictionary. Knowing that this is a dictionary will produce better error messages.

I'm not sure if I can just use { because { denotes a block. I think [{ is better