Parse make expression
I would have like to use the same syntax as swift which is just MyStruct() but there is no way for the syntax analyser to know if this is initialising a MyStruct object or calling a MyStruct function. It get's complicated when arrays and dictionaries get involved. It's probably best to just use a keyword.

"make" instead of "new" because structs are stored on the stack. If I decide to add classes, they'll be stored on the heap. Although, many languages use "new". Hmm...