Add PrecAssoc tables for unary operators
I haven't included the ?: operator

I've read a things about unary operators. OpStack has to differentiate between unary and binary -, and pre and post ++. If I continue using strings, I could do this: "-a", "++a", "a++". I strings are the best option for now. Although, I could make sure that the ast::BinOp, ast::AssignOp, ast::UnOp enums are on separate ranges and store integers in the OpStack instead. popOper already has an if-else chain that checks if the operator is binary, assignment, unary.