Finish rewriting expression parser
I'm using recursive descent everywhere and then all of the sudden, I decided to use shunting yards for expressions. It almost seems silly! I've rewritten it using recursive descent. There is a fair bit of duplication so I'll try to refactor it a bit. Also, much of the expression parser is untested so I'll have to write some tests for every operator and make sure all of the precedence and associativity rules are being followed.

Other than that, I think the syntax analyser is pretty much done!