I THINK I KNOW WHAT I'M DOING NOW
A struct scope is different from a function scope which is different from an enum scope. Insertions and lookups must be handled differently for each type of scope. The struct scope requires additional information for insertions and lookups. For each scope class, there is an insert class and a lookup class. There is also an insert interface and a lookup interface that can be used to hide the details of insertion and lookup.

I have only started implementing this as I write this commit message. I think this will solve a whole lot of problems and make a lot of code much cleaner.

This is the breakthrough that I've been waiting for!

Coverage has probably dropped dramatically after this push. Oh well, it'll be back to 90% in a few weeks.