Sort of support nested types
Consider this:

struct Outer {
  struct Inner {
    static let five = 5;
  }
}

I have made the necessary changes to ExprLookup so that Outer.Inner.five is allowed but Outer.Inner is not. The syntax analyser cannot know if Outer.Inner is a type or not. When the syntax analyser is dealing with an Identifier or a MemberIdent, it could be dealing with a type or a value. In some places, such as in function parameters or member variables, the syntax analyser knows that it is dealing with a type and it can use NamedType or NestedType.

I don't like this inconsistency. Some times it knows and sometimes it doesn't. I would really make a lot more sense for the syntax analyser to just deal with identifiers and let the semantic analyser do all of the lookups to figure out what's a type and what's a value. The only downside to this is a that the syntax highlighter won't know if it's dealing with a type or value so it can't highlight accordingly.