MEGA COMMIT
For a couple of weeks, I didn't really have much time for this project. I did a few things here and there. Then I decided that I wanted dramatically simplify the language. So I rewrote large swaths of code. This stopped the project from compiling for quite a while. I didn't want to commit it in that state so I kept working on it.

I just recently got it compiling and now I have a bit more time to work on this. The lex, format and syntax tests are now passing but the semantic tests are not. I figured I should probably commit now because it's been a while.

I have better understanding of how types should be handled in the semantic analyser. Symbols in the symbol table should refer to types in the ast. Resolving definitions should involve making a ast::FuncCall point to the ast::Func that created it.

I'm still having trouble handling builtin types and builtin operators. Builtin operators are giving me the most trouble. I was implementing builtin operators as functions to make operator overloading easier to implement in future. I don't thing that's really the way to go. I think I need a separate data structure that knows what operations can be performed on builtin types.