Don't need sized integers
There are now just 6 fundamental types instead of 14.

Number parsing has had an overhaul. It's still not that great but it's better than it was before. Suffixes have been introduced to make the type of a number literal obvious. For example, 'A' == 65c && 3000 == 3e3s.

Xcode 10 just got std::variant so I'm using that instead of a tagged union for NumberVariant.