Import a type across modules
MODULES!!!

Implementing this feature has made me realise that a lot of the code needs a lot of the context. Almost everything needs the logger. A lot of things need the scope manager. Type lookup, expression lookup, symbol inserter, builtins, modules. Everything needs everything.

I thought that instead of passing a bajillion individual objects using granular dependency injection, I could just pass in one object that holds references to all of the context.

I saw this in a snippet of a C++ compiler. There were a bunch of functions that just took a pointer to a context struct as a parameter. The context struct had everything in it. This would work well for semantic analysis because semantic analysis is so state-full. Syntax analysis already has this with the ParseTokens object.