Try to get the type of an array literal
We need to construct an ast::ArrayType and have it point to the element type. ast::ArrayType needs a unique pointer to the element type. getExprType returns a raw pointer. The raw pointer might point a builtin type or to the definition of a struct.

I've gotten away with using unique pointers and raw pointers for so long but I think it might be time to switch over to a reference counted pointer. I like the idea of boost::intrusive_ptr but I don't want to depend on boost. I think I'll implement it myself.