Remove an #if
I surrounded some tests with #if 0 /*tests*/ #endif to disable them. My compiler warns be that "Embedding a directive within macro arguments has undefined behavior" but it works on my compiler.

I forgot to remove the #if 0. clang doesn't run half of the tests. gcc produces an executable that crashes.