Deal with nested functions properly
I added heaps of tests to make sure that function lookup is right. When looking up a function, that function can be in the current scope of the parent scope. Variable lookup is slightly different. Nested functions don't capture like a lambda would. Nested functions are essentially anonymous global functions.

Looking up a function for calling is the same as looking up for calling is the same as looking up for a function pointer.

Error messages have improved slightly. We now get "Unreachable variable" instead of "Undefined symbol".