Fix the order of destruction of AST and symbol table
The symbols in the scopes hold retain_ptrs to items in the AST that aren't managed by retain_ptrs (e.g. function parameters). The unmanaged objects are destroyed regardless of their reference count. Managed pointers to these objects are stored in symbols. The retain_ptr will access a pointer to deallocated memory. To prevent this. The scopes must be destroyed before the AST.

I changed the order of the members in sym::Module so that scopes is destroyed before decls.

What puzzles me is that Apple Clang didn't produce an executable that crashed but GCC and Clang did