Setup project for code generation
I'm going to generate C++ code. To test that the C++ code is valid, I'm going to compile it. If there are any warnings, the test has failed. The CXX environment variable is defined on travis. Also travis is running GCC and Clang on Linux and Mac so if it works on my Mac, it should work on travis.