Nested lambda capture
I spent all of yesterday afternoon and all of this morning trying to get this working. But it finally works!

I can't believe my compiler can stomach that Nested nested lambda test! But the return type of makeAdd is difficult to get right. That's why I'm going to implement return type deduction this afternoon.