Test nested for loop
These tests produce almost identical LLVM IR to Clang when compiling the equivalent C. Also, the LLVM optimiser is able to remove the loops and produce a multiply instruction but it's a little bit strange.

The loops optimise down to LLVM IR that looks like this:

if (a == 0 || b == 0) {
  return 0;
} else {
  return a * b;
}

I don't really understand why the optimiser would do that. If either a or b is 0, their product is 0. This sort of thing might make sense if I was dealing the floats instead.