Fix ternary conditional
Previously, both branches of the ternary conditional were evaluated and then one of them was selected based on the condition.

Now, the ternary conditional acts more like an if statement. The true branch is only evaluated if the condition is true. LLVM will still optimise this down to a select instruction if possible.