Introduce value/type categories
The constructors and destructors of lvalues and prvalues are called in all the right places in both Stela and the C++ binding. There's still a hell of a lot to do. xvalues are handled in some places but not others.

We still need to call the destructors of materialised xvalues at the end of an expression. xvalues that are the result of accessing a materialised lvalue don't need their destructors called because they will be destroyed when the parent (materialized) object is destroyed.

I've realised that taking advantage of trivially relocatability is very difficult. It makes me think that I've made a mistake in the constructors or something. I'm probably going to remove the trivial relocation checks and just treat the builtin types as nontrivial types. The optimiser will inline constructors and might be able to elide them in some places. The optimiser knows when optimisations are safe.

In many places, I'm passing the equivalent of array.get() to array constructors in Stela. I'm thinking of removing this because it makes generated code a little hard to read. The optimiser will inline and it won't matter what the function signature is.

Rather than trying to generate fast code, I should try to generate code that is easy to optimise. Give the optimiser nice, clean code and let it do it's thing.

"Premature optimisation is the root of all evil" - Donald Knuth