Change special function signatures
Arrays are treated like nontrivial structs. This results in really clean generated code. The optimised code with the inliner disabled is beautiful! Except, I've noticed that I'm missing out on some optimisations with prvalues.

The C++ standard says that prvalues have a result object. I've only realised what this meant when I looked at some code generated by clang. prvalues are constructed in-place. This means that the code that generates prvalue expressions needs to known where the prvalue is going.